+++ test/utils.dart
@@ -10,22 +10,38 @@ Random _generator() {
 }
 
 /// Generate a list of random numbers of size [length]
-Uint8List randomNumbers(
+List<int> randomNumbers(
   int length, {
   int start = 0,
   int stop = 0xFFFFFFFF,
 }) {
   var random = _generator();
-  var data = Uint8List(length);
-  for (int i = 0; i < data.length; i++) {
-    data[i] = random.nextInt(stop - start + 1) + start;
-  }
-  return data;
+  return List<int>.generate(
+    length,
+    (_) => random.nextInt(stop - start + 1) + start,
+  );
 }
 
 /// Generate a list of random 8-bit numbers of size [length]
 Uint8List randomBytes(int length) {
-  return randomNumbers(length, stop: 0xFF);
+  var random = _generator();
+  var data = Uint8List(length);
+  for (int i = 0; i < data.length; i++) {
+    data[i] = random.nextInt(256);
+  }
+  return data;
+}
+
+/// Generate a list of random valid Unicode code points of size [length],
+/// spanning the full range up to U+10FFFF, excluding surrogates.
+List<int> randomCodePoints(int length) {
+  var random = _generator();
+  return List<int>.generate(length, (_) {
+    for (;;) {
+      var c = random.nextInt(0x110000);
+      if (c < 0xD800 || c > 0xDFFF) return c;
+    }
+  });
 }
 
 /// Fill the [buffer] with random numbers